fix failing test for itemlist, move tests into correct package
