+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/utils/items/ItemListTest.java
@@ -1,6 +1,5 @@
-package de.saxsys.mvvmfx.utils.newitemlist;
+package de.saxsys.mvvmfx.utils.items;
 
-import de.saxsys.mvvmfx.utils.items.ItemList;
 import javafx.collections.ListChangeListener;
 import org.junit.Before;
 import org.junit.Test;
