+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/utils/items/ListViewItemListTest.java
@@ -1,7 +1,6 @@
-package de.saxsys.mvvmfx.utils.newitemlist;
+package de.saxsys.mvvmfx.utils.items;
 
 import de.saxsys.mvvmfx.testingutils.jfxrunner.JfxRunner;
-import de.saxsys.mvvmfx.utils.items.ItemList;
 import javafx.scene.control.ListView;
 import org.junit.Before;
 import org.junit.Test;
